refactor(activerecord): merge QueryCacheStore into abstract/query-cache Store
Rails has one `ActiveRecord::ConnectionAdapters::QueryCache::Store`.
We had split it into two classes: an LRU-bearing `QueryCacheStore`
in `query-cache.ts` and a thin `Store extends QueryCacheStore`
wrapper in `abstract/query-cache.ts` that only added `isDirties()`.
The split existed because `QueryCacheAdapter` (our TS-only wrapper)
lived in `query-cache.ts` and wanted the LRU class without importing
from `abstract/`.

Folds the LRU state + methods into the `Store` class in
`abstract/query-cache.ts` (its Rails-canonical home), with
`isDirties()` preserved. Deletes the outer `QueryCacheStore`;
`QueryCacheAdapter` now imports `Store` from abstract/. The
package index re-exports it as `QueryCacheStore` so the public
surface stays disambiguated from generic `Store` symbols.

With one class, the extends chain is gone — no TS_ROOT_INTERMEDIATE
entry needed for this pair.
