+++ packages/activerecord/src/connection-adapters/abstract/query-cache.ts
@@ -1,15 +1,74 @@
 import { Notifications } from "@blazetrails/activesupport";
-import { QueryCacheStore } from "../../query-cache.js";
 import type { DatabaseStatementsHost } from "./database-statements.js";
 
 const DEFAULT_SIZE = 100;
+const DEFAULT_MAX_SIZE = 100;
 
 /**
+ * LRU cache store for query results.
+ *
  * Mirrors: ActiveRecord::ConnectionAdapters::QueryCache::Store
  */
-export class Store extends QueryCacheStore {
+export class Store {
+  private _map = new Map<string, Record<string, unknown>[]>();
+  private _maxSize: number;
+  enabled = false;
+  dirties = true;
+
+  constructor(maxSize: number = DEFAULT_MAX_SIZE) {
+    this._maxSize = maxSize;
+  }
+
+  get size(): number {
+    return this._map.size;
+  }
+
+  get empty(): boolean {
+    return this._map.size === 0;
+  }
+
   isDirties(): boolean {
     return this.dirties;
+  }
+
+  get(key: string): Record<string, unknown>[] | undefined {
+    if (!this.enabled) return undefined;
+    const entry = this._map.get(key);
+    if (entry) {
+      // Move to end (LRU)
+      this._map.delete(key);
+      this._map.set(key, entry);
+    }
+    return entry;
+  }
+
+  computeIfAbsent(
+    key: string,
+    compute: () => Promise<Record<string, unknown>[]>,
+  ): Promise<Record<string, unknown>[]> {
+    if (!this.enabled) return compute();
+
+    const cached = this.get(key);
+    if (cached !== undefined) {
+      return Promise.resolve(cached.map((row) => ({ ...row })));
+    }
+
+    return compute().then((result) => {
+      if (this._maxSize <= 0) {
+        // maxSize of 0 or negative disables caching — return without storing
+        return result;
+      }
+      if (this._map.size >= this._maxSize) {
+        const firstKey = this._map.keys().next().value;
+        if (firstKey !== undefined) this._map.delete(firstKey);
+      }
+      this._map.set(key, result);
+      return result.map((row) => ({ ...row }));
+    });
+  }
+
+  clear(): void {
+    this._map.clear();
   }
 }
