+++ packages/activerecord/src/query-cache.test.ts
@@ -1,7 +1,8 @@
 import { describe, it, expect } from "vitest";
 import { Base } from "./index.js";
 import { createTestAdapter } from "./test-adapter.js";
-import { QueryCache, QueryCacheAdapter, QueryCacheStore } from "./query-cache.js";
+import { QueryCache, QueryCacheAdapter } from "./query-cache.js";
+import { Store } from "./connection-adapters/abstract/query-cache.js";
 
 function setup() {
   const inner = createTestAdapter();
@@ -383,7 +384,7 @@ describe("QueryCacheExpiryTest", () => {
   });
 
   it("enable disable", async () => {
-    const store = new QueryCacheStore();
+    const store = new Store();
     expect(store.enabled).toBe(false);
     store.enabled = true;
     expect(store.enabled).toBe(true);
@@ -405,7 +406,7 @@ describe("QueryCacheExpiryTest", () => {
   });
 
   it("query cache lru eviction", async () => {
-    const store = new QueryCacheStore(3);
+    const store = new Store(3);
     store.enabled = true;
     for (let i = 0; i < 5; i++) {
       await store.computeIfAbsent(`query_${i}`, async () => [{ val: i }]);
